+++ src/includes/class-i18n.php
@@ -40,11 +40,9 @@ class I18n {
 		load_plugin_textdomain(
 			'bh-wp-plugins-page',
 			false,
-			dirname( dirname( plugin_basename( __FILE__ ) ) ) . '/languages/'
+            plugin_basename( dirname( __FILE__, 2 ) ) . '/languages/'
 		);
 
 	}
-
-
 
 }