+++ CHANGELOG.md
@@ -1,2 +1,9 @@
 # Changelog
 
+## 1.0.7 2012-09-15
+
+* Fix bug when a plugin was providing an empty link ( just '' !).
+
+## 1.0.6 2021-09-13
+
+* Pass all plugin data to apply_filters (some plugins rely on it and were logging errors)